latest updates

+++ viikko3/PDO-harjoitus/operations/selectData.php
@@ -19,8 +19,14 @@ try {
             <td><img src="<?php echo $SITE_URL; ?>/uploads/<?php echo $row['filename']; ?>"
                      alt="<?php echo $row['title']; ?>"></td>
             <td>
-                <button class="modify-button" data-media_id="<?php echo $row['media_id']; ?>">Modify</button>
-                <a href="<?php echo $SITE_URL; ?>/operations/deleteData.php?media_id=<?php echo $row['media_id']; ?>">Delete</a>
+                <?php if ($_SESSION['user']['user_id'] == $row['user_id']
+                    || $_SESSION['user']['user_level_id'] == 1): ?>
+                    <button class="modify-button" data-media_id="<?php echo $row['media_id']; ?>">Modify</button>
+                    <!--suppress HtmlUnknownTarget -->
+                    <a href="<?php echo $SITE_URL; ?>/operations/deleteData.php?media_id=<?php echo $row['media_id']; ?>">Delete</a>
+                <?php else: ?>
+                    -
+                <?php endif; ?>
             </td>
         </tr>
     <?php
